refactored MMDB related items
