+++ src/networking/manage_packets.rs
@@ -6,7 +6,9 @@ use dns_lookup::lookup_addr;
 use etherparse::{Ethernet2Header, IpHeader, PacketHeaders, TransportHeader};
 use pcap::{Active, Address, Capture, Device};
 
-use crate::countries::country_utils::get_country;
+use crate::mmdb::asn::get_asn;
+use crate::mmdb::country::get_country;
+use crate::mmdb::types::mmdb_reader::MmdbReader;
 use crate::networking::types::address_port_pair::AddressPortPair;
 use crate::networking::types::app_protocol::from_port_to_application_protocol;
 use crate::networking::types::data_info_host::DataInfoHost;
@@ -16,7 +18,6 @@ use crate::networking::types::info_address_port_pair::InfoAddressPortPair;
 use crate::networking::types::my_device::MyDevice;
 use crate::networking::types::traffic_direction::TrafficDirection;
 use crate::networking::types::traffic_type::TrafficType;
-use crate::utils::asn::{asn, MmdbReader};
 use crate::utils::formatted_strings::get_domain_from_r_dns;
 use crate::IpVersion::{IPv4, IPv6};
 use crate::{AppProtocol, InfoTraffic, IpVersion, TransProtocol};
@@ -262,7 +263,7 @@ pub fn reverse_dns_lookup(
     );
     let is_local = is_local_connection(&address_to_lookup, &my_interface_addresses);
     let country = get_country(&address_to_lookup, country_db_reader);
-    let asn = asn(&address_to_lookup, asn_db_reader);
+    let asn = get_asn(&address_to_lookup, asn_db_reader);
     let r_dns = if let Ok(result) = lookup_result {
         if result.is_empty() {
             address_to_lookup.clone()
