+++ src/utils/mod.rs
@@ -1,3 +1,2 @@
-pub mod asn;
 pub mod formatted_strings;
 pub mod types;
